Move the compression detection functions to the public header
In response to https://github.com/facebook/osquery/pull/1284, we would
like to add a test to ensure that the RocksDB we're building with was
compiled with Snappy support.

If you all would prefer a different approach with this, happy to oblige.

I successfully built and ran the tests for this on my local machine.
