+++ db/compaction_job_stats_test.cc
@@ -30,6 +30,7 @@
 #include "port/stack_trace.h"
 #include "rocksdb/cache.h"
 #include "rocksdb/compaction_filter.h"
+#include "rocksdb/compression.h"
 #include "rocksdb/db.h"
 #include "rocksdb/env.h"
 #include "rocksdb/experimental.h"
